requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-anymail==13.1
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.22.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycryptodome==3.23.0
PyJWT==2.10.1
qrcode==8.2
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
